oidn

Struct RayTracing

pub struct RayTracing<'a> { /* private fields */ }

A generic ray tracing denoising filter for denoising images produces with Monte Carlo ray tracing methods such as path tracing.

Implementations

impl<'a> RayTracing<'a>

pub fn new(device: &'a Device) -> RayTracing<'a>

pub fn filter_quality(&mut self, quality: Quality) -> &mut RayTracing<'a>

Sets the quality of the output, the default is high.

Balanced lowers the precision, if possible, however some devices will not support this and so the result (and performance) will stay the same as high. Balanced is recommended for realtime usages.

pub fn albedo_normal( &mut self, albedo: &[f32], normal: &[f32], ) -> &mut RayTracing<'a>

Set input auxiliary images containing the albedo and normals.

Albedo must have three channels per pixel with values in [0, 1]. Normal must contain the shading normal as three channels per pixel world-space or view-space vectors with arbitrary length, values in [-1, 1].

Panics

if resource creation fails

pub fn albedo(&mut self, albedo: &[f32]) -> &mut RayTracing<'a>

Set an input auxiliary image containing the albedo per pixel (three channels, values in [0, 1]).

Panics

if resource creation fails

pub fn albedo_normal_buffer( &mut self, albedo: Buffer, normal: Buffer, ) -> Option<&mut RayTracing<'a>>

Set input auxiliary buffer containing the albedo and normals.

Albedo buffer must have three channels per pixel with values in [0, 1]. Normal must contain the shading normal as three channels per pixel world-space or view-space vectors with arbitrary length, values in [-1, 1].

This function is the same as RayTracing::albedo_normal but takes buffers instead

Returns None if either buffer was not created by this device

pub fn albedo_buffer(&mut self, albedo: Buffer) -> Option<&mut RayTracing<'a>>

Set an input auxiliary buffer containing the albedo per pixel (three channels, values in [0, 1]).

This function is the same as RayTracing::albedo but takes buffers instead

Returns None if albedo buffer was not created by this device

pub fn hdr(&mut self, hdr: bool) -> &mut RayTracing<'a>

Set whether the color is HDR.

pub fn hdr_scale(&mut self, hdr_scale: f32) -> &mut RayTracing<'a>

👎Deprecated since 1.3.1:
Please use RayTracing::input_scale instead

pub fn input_scale(&mut self, input_scale: f32) -> &mut RayTracing<'a>

Sets a scale to apply to input values before filtering, without scaling the output too.

This can be used to map color or auxiliary feature values to the expected range. E.g. for mapping HDR values to physical units (which affects the quality of the output but not the range of the output values). If not set, the scale is computed implicitly for HDR images or set to 1 otherwise

pub fn srgb(&mut self, srgb: bool) -> &mut RayTracing<'a>

Set whether the color is encoded with the sRGB (or 2.2 gamma) curve (LDR only) or is linear.

The output will be encoded with the same curve.

pub fn clean_aux(&mut self, clean_aux: bool) -> &mut RayTracing<'a>

Set whether the auxiliary feature (albedo, normal) images are noise-free.

Recommended for highest quality but should not be enabled for noisy auxiliary images to avoid residual noise.

pub fn image_dimensions( &mut self, width: usize, height: usize, ) -> &mut RayTracing<'a>

sets the dimensions of the denoising image, if new width * new height does not equal old width * old height

pub fn filter(&self, color: &[f32], output: &mut [f32]) -> Result<(), Error>

pub fn filter_buffer( &self, color: &Buffer, output: &Buffer, ) -> Result<(), Error>

pub fn filter_in_place(&self, color: &mut [f32]) -> Result<(), Error>

pub fn filter_in_place_buffer(&self, color: &Buffer) -> Result<(), Error>